+++ src/test/java/io/hhplus/tdd/point/service/unit/PointHistorySearchTest.java
@@ -1,7 +1,8 @@
-package io.hhplus.tdd.point.service;
+package io.hhplus.tdd.point.service.unit;
 
 import io.hhplus.tdd.point.PointHistory;
 import io.hhplus.tdd.point.repository.PointHistoryRepository;
+import io.hhplus.tdd.point.service.PointService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
