+++ src/test/java/io/hhplus/tdd/point/service/unit/PointSearchTest.java
@@ -1,7 +1,8 @@
-package io.hhplus.tdd.point.service;
+package io.hhplus.tdd.point.service.unit;
 
 import io.hhplus.tdd.point.UserPoint;
 import io.hhplus.tdd.point.repository.UserPointRepository;
+import io.hhplus.tdd.point.service.PointService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
